+++ CHANGELOG.md
@@ -15,14 +15,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+### Changed
+### Fixed
+### Removed
+### Deprecated
+### Security
+
+
+## [1.12.0] - 2026-08-02
+### Added
 - Decode new pump telemetry. Cmd 0x3B is extended with two extra bytes of payload that encode the pump's power consumption (fixes #81).
 ### Changed
 - Refactored channel types to use a strongly typed `channel_type_t` enum across decoder and state (fixes #84).
 ### Fixed
 - Set pump speed to 0 when the Filter channel is not active (fixes #82).
-### Removed
-### Deprecated
-### Security
 
 ## [1.11.2] - 2026-07-27
 ### Fixed